dev-requirements.txt
dev.asgiref==3.8.1
Django==5.0
psycopg==3.2.1
psycopg-binary==3.2.1
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
django-ninja==1.2.1